Fixed vcxproj hash error. Finalised radiation and conduction physics. Changed data types to improve conservation tests.

+++ include/physics.h
@@ -5,10 +5,10 @@
 class Physics {
 
 public:
-	void DeepSpaceHeatTransfer(std::vector <ThermalBlocks > * block, float dt, float transferspeed, float sigma, int emissivety);
+	void DeepSpaceHeatTransfer(std::vector <ThermalBlocks > * block, float dt, float transferspeed, double sigma, double emissivety);
 	//float HeatTransfer();
 	void AddConduction(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, float transferspeed, std::vector<float>& tempsToadd, BlockManager& blockUser);
 	void AddRadiation(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, float transferspeed, float sigma, std::vector<float>& tempsToadd, BlockManager& blockUser);
 	void GetRGB(ThermalBlocks* blocks);;
-	void getCoolingConstant(float sigma, std::vector<ThermalBlocks>& blocksinuse);
+	void getCoolingConstant(double sigma, std::vector<ThermalBlocks>& blocksinuse);
 };
